avoid circular dep on import
